+++ src/dawn/core/io/InputStream.cpp
@@ -10,7 +10,7 @@ namespace dw {
 InputStream::InputStream() : position_(0), size_(0) {
 }
 
-InputStream::InputStream(u64 size) : position_(0), size_(size) {
+InputStream::InputStream(usize size) : position_(0), size_(size) {
 }
 
 InputStream::~InputStream() {
@@ -33,11 +33,11 @@ bool InputStream::eof() const {
     return position_ >= size_;
 }
 
-u64 InputStream::position() const {
+usize InputStream::position() const {
     return position_;
 }
 
-u64 InputStream::size() const {
+usize InputStream::size() const {
     return size_;
 }
 }  // namespace dw
