+++ src/dawn/renderer/SceneGraph.h
@@ -37,7 +37,7 @@ public:
     Frame* addFrame(SystemNode* frame_node);
     void removeFrame(Frame* frame);
     Frame* frame(int i);
-    uint frameCount() const;
+    usize frameCount() const;
 
     // Callbacks. All take delta time (dt) as first parameter.
     Function<void(float)> preRenderSceneCallback;
